requirements-local.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.6
easydict==1.13
face-alignment==1.4.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.3
numba==0.62.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnx==1.19.1
onnxruntime-gpu==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
prettytable==3.16.0
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.3
six==1.17.0
stringzilla==4.2.1
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14

